Clear the warning under the 32-bit system and use uintptr_t instead of uint64_t as the string pointer type.

+++ ulog.h
@@ -95,7 +95,7 @@
                _TYPE_CMP(token, char[]) || _TYPE_CMP(token, const char[]) ||  \
                _TYPE_CMP(token, unsigned char[]) ||                           \
                _TYPE_CMP(token, const unsigned char[])) {                     \
-      const char *token_value = (const char *)(uint64_t)(token);              \
+      const char *token_value = (const char *)(uintptr_t)(token);             \
       _LOG_DEBUG(_LOG_TOKEN_FORMAT("(char*)[%" PRIu32 "]", "%s"),             \
                  (uint32_t)strlen(token_value), #token, token_value);         \
     } else if (_TYPE_CMP(token, void *) || _TYPE_CMP(token, short *) ||       \
